pyproject.toml
[project]
name = "mcp-servers-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-mcp-adapters>=0.0.9",
]
